Increase the min doc count to 5
