+++ remote_vector_index_builder/core/common/models/__init__.py
@@ -6,28 +6,8 @@
 # compatible open source license.
 
 from .index_build_parameters import SpaceType
-
-from .index_builder.cagra_graph_build_algo import CagraGraphBuildAlgo
-
-
 from .index_build_parameters import IndexBuildParameters
 from .vectors_dataset import VectorsDataset
-from .index_builder.response.faiss_gpu_build_index_output import (
-    FaissGpuBuildIndexOutput,
-)
-from .index_builder.response.faiss_cpu_build_index_output import (
-    FaissCpuBuildIndexOutput,
-)
-from .index_builder.faiss_gpu_index_builder import FaissGPUIndexBuilder
-from .index_builder.faiss_cpu_index_builder import FaissCPUIndexBuilder
 
-__all__ = [
-    "SpaceType",
-    "CagraGraphBuildAlgo",
-    "IndexBuildParameters",
-    "VectorsDataset",
-    "FaissGpuBuildIndexOutput",
-    "FaissCpuBuildIndexOutput",
-    "FaissGPUIndexBuilder",
-    "FaissCPUIndexBuilder",
-]
+
+__all__ = ["SpaceType", "IndexBuildParameters", "VectorsDataset"]
