+++ remote_vector_index_builder/core/common/models/index_builder/__init__.py
@@ -1,0 +1,24 @@
+# Copyright OpenSearch Contributors
+# SPDX-License-Identifier: Apache-2.0
+#
+# The OpenSearch Contributors require contributions made to
+# this file be licensed under the Apache-2.0 license or a
+# compatible open source license.
+
+from .cagra_graph_build_algo import CagraGraphBuildAlgo
+from .response.faiss_gpu_build_index_output import (
+    FaissGpuBuildIndexOutput,
+)
+from .response.faiss_cpu_build_index_output import (
+    FaissCpuBuildIndexOutput,
+)
+from .faiss_gpu_index_builder import FaissGPUIndexBuilder
+from .faiss_cpu_index_builder import FaissCPUIndexBuilder
+
+__all__ = [
+    "CagraGraphBuildAlgo",
+    "FaissGpuBuildIndexOutput",
+    "FaissCpuBuildIndexOutput",
+    "FaissGPUIndexBuilder",
+    "FaissCPUIndexBuilder",
+]
