+++ remote_vector_index_builder/core/common/models/index_builder/faiss_gpu_index_builder.py
@@ -10,6 +10,9 @@ from abc import ABC, abstractmethod
 from core.common.models import (
     SpaceType,
     VectorsDataset,
+)
+
+from core.common.models.index_builder import (
     FaissGpuBuildIndexOutput,
 )
