+++ remote_vector_index_builder/core/index_builder/index_builder_utils.py
@@ -10,7 +10,7 @@ import math
 import faiss
 
 
-from core.common.models.index_build_parameters import (
+from core.common.models import (
     SpaceType,
 )
