+++ test_remote_vector_index_builder/test_core/common/models/index_builder/faiss/test_faiss_gpu_index_caga_builder.py
@@ -7,7 +7,14 @@
 
 import faiss
 import pytest
-from core.common.models import CagraGraphBuildAlgo, SpaceType, FaissGpuBuildIndexOutput
+
+from core.common.models import SpaceType
+
+from core.common.models.index_builder import (
+    CagraGraphBuildAlgo,
+    FaissGpuBuildIndexOutput,
+)
+
 from core.common.models.index_builder.faiss import (
     IVFPQBuildCagraConfig,
     IVFPQSearchCagraConfig,
